requirements.txt
alembic==1.3.0
aniso8601==8.0.0
cffi==1.13.1
Click==7.0
cryptography==2.8
Flask==1.1.1
Flask-Cors==3.0.8
Flask-HTTPAuth==3.3.0
Flask-Migrate==2.5.2
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.1
freeze==1.0.10
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
pycparser==2.19
pymongo==3.9.0
PyMySQL==0.9.3
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.3
six==1.12.0
SQLAlchemy==1.3.10
Werkzeug==0.16.0
